requirements.txt
Flask==3.0.3
boto3==1.34.131
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0